+++ numcpp/opencv.h
@@ -32,7 +32,7 @@ const cv::Mat to_cv_mat(const array_t<T> &array)
 
 inline base_array_t from_cv_mat(const cv::Mat &cv_mat)
 {
-	base_array_t result(cv_mat.elemSize());
+	base_array_t result((int)cv_mat.elemSize());
 	result.setSize<heap_allocator>(cv_mat.cols, cv_mat.rows);
 
 	// TODO: Do not copy
